fix: allow to use full name in default value reference
For schemas where a field with default value has the same name as its
type, avrohugger will produce code that does not compile, because the
compiler does not know if the reference is related to the field or the
imported type.

This is a small fix which just checks for this condition and uses the
full name in the reference in that case. Currently this only done for
the `specific` generation to provide a minimal invasive PR, but should
also apply to the others.